feat(applications): restructure detail page from 7 tabs to 4
The application detail page had grown to 7 tabs (Overview, Services,
Routing, Pool, Monitoring, History, Configuration) with the same
information scattered across them — routing shown on its own tab and inside
Configuration, networks/volumes in three places, history both as a tab and
on Overview, and Pool as a near-empty top-level tab for most apps.

Collapse to 4 intent-based tabs and promote lifecycle actions to the header:

- Overview — status, connect, networks, details, recent activity
- Services — Services + Routing + Pool merged into one "what's running" view
  (routing endpoints and pool instances render only when present)
- Configuration — the settings-rail edit form
- Activity — Monitoring + History merged into one "over time" view

Deploy / Stop / Delete move from the Overview "Actions" card into the page
header, so they're reachable from every tab and the two duplicate deploy
buttons collapse to one.

- Routing/Pool/Monitoring/History tab bodies extracted into reusable
  _components (routing-endpoints, pool-section, monitoring-section,
  history-section); the standalone page.tsx files are removed.
- Legacy /routing, /pool, /monitoring, /history paths redirect to the
  merged tabs so existing bookmarks/links still resolve.
- Network-membership editing (ConnectedNetworksCard) stays on Overview: it
  mutates the live stack, not the template draft the Configuration form edits.
